restructuration fichier avec navbar footer dans router + error page
