+++ src/index.js
@@ -6,17 +6,21 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import Home from './pages/Home/Home';
 import About from './pages/About/About';
 import Error from './pages/Error/Error';
+import NavBar from './components/NavBar/NavBar';
+import Footer from './components/Footer/Footer';
 /* import Carousel from './pages/Carousel/'; */
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <Router>
+      <NavBar />
       <Routes>
         <Route path="/" element={<Home />}></Route>
         <Route path="/about" element={<About />}></Route>
         <Route path="*" element={<Error />}></Route>
       </Routes>
+      <Footer />
     </Router>
   </React.StrictMode>
 );
