+++ src/pages/Home/Home.jsx
@@ -1,16 +1,11 @@
-import NavBar from '../../components/NavBar/NavBar';
 import Banner from '../../components/Banner/Banner';
-import Footer from '../../components/Footer/Footer';
-//import '../App.css';
 
-function App() {
+function Home() {
   return (
-    <div className="App">
-      <NavBar />
+    <main>
       <Banner />
-      <Footer />
-    </div>
+    </main>
   );
 }
 
-export default App;
+export default Home;
